nicer log output

+++ src/mavsdk/core/mavlink_parameter_sender.cpp
@@ -1029,7 +1029,7 @@ bool MavlinkParameterSender::source_matches(uint8_t source_sys_id,uint8_t source
     const bool matches=source_sys_id==_sender.get_system_id() && source_comp_id== _target_component_id;
     if(!matches){
         LogDebug()<<"Skipping message coming from {"<<(int)source_sys_id<<":"<<(int)source_comp_id<<"} "
-                   <<" since we talk to {"<<_sender.get_system_id()<<":"<<_target_component_id<<"}";
+                   <<" since we talk to {"<<(int)_sender.get_system_id()<<":"<<(int)_target_component_id<<"}";
     }
     return matches;
 }
